[aptos-workspac-server] graceful shutdown + modularize code
