+++ aptos-move/aptos-workspace-server/src/services/mod.rs
@@ -1,0 +1,6 @@
+// Copyright (c) Aptos Foundation
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod faucet;
+pub mod node;
+pub mod postgres;
